requirements.txt
beautifulsoup4==4.8.0
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-registration==2.4.1
django-tinymce==2.8.0
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
soupsieve==1.9.3
whitenoise==4.1.3
